Move methods and split unit tests.
