+++ src/main/java/org/fundacionjala/coding/manuel/encryptionalgorithms/Vigenere.java
@@ -1,9 +1,12 @@
 package org.fundacionjala.coding.manuel.encryptionalgorithms;
 
+import java.util.StringJoiner;
+
 /**
  * Class for Vigenere algorithm.
  */
 public class Vigenere extends Crypto {
+    private static final int VALUE = 64;
 
     /**
      * Method to encode text using Vigenere algorithm.
@@ -29,5 +32,29 @@ public class Vigenere extends Crypto {
             return "";
         }
         return vigenereCoreCall(objects[0].toString(), objects[1].toString(), -1);
+    }
+
+    /**
+     * Method to encode or decode text with Vigenere depending on direction parameter.
+     * @param text to encode or decode.
+     * @param key to encode or decode.
+     * @param direction to move right or left.
+     * @return the text encrypted.
+     */
+    public String vigenereCoreCall(String text, String key, int direction) {
+        int j = 0;
+        StringJoiner encriptedText = new StringJoiner(" ");
+        for (String word : text.split(" ")) {
+            StringBuilder encriptedWord = new StringBuilder();
+            for (int i = 0; i < word.length(); i++, j++) {
+                if (j == key.length()) {
+                    j = 0;
+                }
+                int newValue = determineNumberMovements(word.charAt(i), (key.charAt(j) - VALUE) * direction);
+                encriptedWord.append(String.valueOf(Character.toChars(newValue)));
+            }
+            encriptedText.add(encriptedWord);
+        }
+        return encriptedText.toString();
     }
 }
