+++ src/test/java/org/fundacionjala/coding/manuel/encryptionalgorithms/CesarTest.java
@@ -1,29 +1,19 @@
-package org.fundacionjala.coding.manuel;
+package org.fundacionjala.coding.manuel.encryptionalgorithms;
 
-import org.fundacionjala.coding.manuel.encryptionalgorithms.Cesar;
-import org.fundacionjala.coding.manuel.encryptionalgorithms.Crypto;
-import org.fundacionjala.coding.manuel.encryptionalgorithms.Vigenere;
+import static org.junit.Assert.assertEquals;
 import org.junit.Before;
 import org.junit.Test;
 
-import static org.junit.Assert.assertEquals;
-
 /**
- * Class to test CesarVigenereCrypto.
+ * Class to test Cesar class.
  */
-public class CesarVigenereCryptoTest {
+public class CesarTest {
 
     /**
      * Object of the class Cesar.
      */
     private Crypto cesarCrypto;
 
-    /**
-     * Object of the class Vigenere.
-     */
-    private Crypto vigenereCrypto;
-
-    private static final String VIGENERE_KEY = "LIMON";
     private static final String TEST_WORD = "ZORRO";
 
     /**
@@ -32,7 +22,6 @@ public class CesarVigenereCryptoTest {
     @Before
     public void init() {
         cesarCrypto = new Cesar();
-        vigenereCrypto = new Vigenere();
     }
 
     /**
@@ -98,43 +87,5 @@ public class CesarVigenereCryptoTest {
         final int secondKey = 5;
         assertEquals("HO_LA", cesarCrypto.decode("KR_OD", key));
         assertEquals("ZOR;RO", cesarCrypto.decode("ETW;WT", secondKey));
-    }
-
-    /**
-     * Test Vigenere encodeCesar method.
-     */
-    @Test
-    public void testVigenereEncode() {
-        assertEquals("TXYP", vigenereCrypto.encode("HOLA", VIGENERE_KEY));
-        assertEquals("LXEGC", vigenereCrypto.encode(TEST_WORD, VIGENERE_KEY));
-        assertEquals("TXYP XMUNHCRC", vigenereCrypto.encode("HOLA JALASOFT", VIGENERE_KEY));
-    }
-
-    /**
-     * Test Vigenere encodeCesar method for null and empty texts.
-     */
-    @Test
-    public void testVigenereEncodeNullAndEmpty() {
-        assertEquals("", vigenereCrypto.encode(null, VIGENERE_KEY));
-        assertEquals("", vigenereCrypto.encode("", VIGENERE_KEY));
-    }
-
-    /**
-     * Test Vigenere decodeCesar method.
-     */
-    @Test
-    public void testVigenereDecode() {
-        assertEquals("HOLA", vigenereCrypto.decode("TXYP", VIGENERE_KEY));
-        assertEquals(TEST_WORD, vigenereCrypto.decode("LXEGC", VIGENERE_KEY));
-        assertEquals("HOLA JALASOFT", vigenereCrypto.decode("TXYP XMUNHCRC", VIGENERE_KEY));
-    }
-
-    /**
-     * Test Vigenere encodeCesar method for null and empty texts.
-     */
-    @Test
-    public void testVigenereDecodeNullAndEmpty() {
-        assertEquals("", vigenereCrypto.decode(null, VIGENERE_KEY));
-        assertEquals("", vigenereCrypto.decode("", VIGENERE_KEY));
     }
 }
